requirements.txt
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
ipaddress==1.0.23
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
pycparser==2.22
PyJWT==2.8.0
pymongo==4.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
